Cleanup & Fixes
Cleaned the Project Filters, fixed a bug that was caused because GameState loaded the same textures twice when you started the game twice. 

Removed an unnecessary call from ShakeEffect. 

Removed console messages.

Fixed a little issue in the collision detection checks for when the winner is declared.

Removed Constants and Arena as they are unused.